+++ src/cmd/go/scriptreadme_test.go
@@ -6,7 +6,7 @@ package main_test
 
 import (
 	"bytes"
-	"cmd/go/internal/script"
+	"cmd/internal/script"
 	"flag"
 	"internal/diff"
 	"internal/testenv"
@@ -38,7 +38,7 @@ func checkScriptReadme(t *testing.T, engine *script.Engine, env []string) {
 	args.Conditions = conds.String()
 
 	doc := new(strings.Builder)
-	cmd := testenv.Command(t, testGo, "doc", "cmd/go/internal/script")
+	cmd := testenv.Command(t, testGo, "doc", "cmd/internal/script")
 	cmd.Env = env
 	cmd.Stdout = doc
 	if err := cmd.Run(); err != nil {
