+++ src/cmd/internal/script/cmds.go
