+++ src/cmd/internal/script/cmds_nonunix.go
