+++ src/cmd/internal/script/cmds_unix.go
