+++ src/cmd/internal/script/conds.go
