+++ src/cmd/internal/script/errors.go
