+++ src/cmd/internal/script/state.go
